pyproject.toml
[tool.poetry]
name = "ticketer"
version = "0.1.0"
description = ""
authors = ["RuslanUC <dev_ruslan_uc@protonmail.com>"]
readme = "README.md"
packages = [{include = "ticketer"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
tortoise-orm = {extras = ["asyncmy", "accel"], version = "^0.20.1"}
s3lite = "^0.1.4"
bcrypt = "^4.1.3"
uvicorn = "^0.29.0"
httpx = "^0.27.0"
aerich = "^0.7.2"
pydantic = {extras = ["email"], version = "^2.7.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
